+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+# [0.72.0](https://github.com/lightdash/lightdash/compare/0.71.7...0.72.0) (2022-04-22)
+
+
+### Features
+
+* add description field to create chart modal ([#1811](https://github.com/lightdash/lightdash/issues/1811)) ([ce82334](https://github.com/lightdash/lightdash/commit/ce82334ac00f2efa318e5d3bf68adb7119267f23))
+
+## [0.71.7](https://github.com/lightdash/lightdash/compare/0.71.6...0.71.7) (2022-04-21)
+
+
+### Bug Fixes
+
+* persist pivot layout ([#1834](https://github.com/lightdash/lightdash/issues/1834)) ([69289e0](https://github.com/lightdash/lightdash/commit/69289e06a520a0de3d3aec4114317aa1b8f3acbc))
+
 ## [0.71.6](https://github.com/lightdash/lightdash/compare/0.71.5...0.71.6) (2022-04-21)